requirements.txt
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycryptodome==3.17
PyJWT==2.6.0
pytz==2022.7.1
wincertstore==0.2
